Moved tag model to models file in profiles and added fixtures folder
