+++ profiles/admin.py
@@ -1,2 +1,6 @@
 from django.contrib import admin
+from .models import Profile, Tag
+
+admin.site.register(Profile)
+admin.site.register(Tag)
