+++ profiles/templates/profile.html
@@ -10,14 +10,13 @@
   <!-- Profile Information -->
   <div class="d-flex flex-column flex-sm-row my-3 py-3 justify-content-center">
     <div class="flex-column p-3 text-center">
-        <img src="{{ user.profile.profile_image.url }}" alt="user profile image" class="rounded-circle profile-pic">
+        <img src="{{ user.profile.profile_pic.url }}" alt="user profile image" class="rounded-circle profile-pic">
     </div>
     <div class="d-flex flex-column p-3 text-center">
         <h3>My Details</h3>
         <p>Name: {{ user.first_name }} {{ user.last_name }}</p>
-        <p>Email Address: {{ user.email }}</p>
-        {% if user.profile.location %}
-          <p>Location: {{ user.profile.location }}</p>
+        {% if user.email %}
+          <p>Email Address: {{ user.email }}</p>
         {% endif %}
     </div>
   </div>
